requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
packaging==23.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
